Check if file is empty before deleting it.
If file is empty the delete method will try to delete MEDIA_ROO and raise IOError.

+++ user_clipboard/models.py
@@ -87,4 +87,5 @@ def _handle_deleteing(sender, instance, **kwargs):
     """
     Delete file when Clipboard item was deleted
     """
-    instance.file.delete(save=False)
+    if instance.file:
+        instance.file.delete(save=False)
